fix run_tests.sh

+++ run_tests.sh
@@ -1,7 +1,7 @@
 #!/bin/bash
 IFS=$'\n'
 
-rm output.txt
+[ -e output.txt ] && rm output.txt
 
 ./main scenario1.json >> output.txt
 ./main scenario2.json >> output.txt
